+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Change Log
 Next Release
 ------------
 - Feature: ``ObjectView`` and ``ObjectMixin`` for rendering single objects
+- Feature: ``request.view_args`` can now be accessed as view instance
+  attributes for accessing in other methods
 
 2014.04.24.2
 ------------