+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Next Release
 - Feature: ``ObjectView`` and ``ObjectMixin`` for rendering single objects
 - Feature: ``request.view_args`` can now be accessed as view instance
   attributes for accessing in other methods
+- Feature: Ability to customise context name for ``objects`` / ``object``
+  within SQLAlchemy read views.
 
 2014.04.24.2
 ------------